move load_dragon_profile
