+++ identity-wallet/src/state/reducers/dynamic_dev_profile.rs
@@ -3,16 +3,73 @@ use crate::{
     command,
     error::AppError,
     state::{
-        actions::{ConnectionAccepted, CreateNew, CredentialOffersSelected, CredentialsSelected, QrCodeScanned, Reset},
+        actions::{ConnectionAccepted, CreateNew, CredentialOffersSelected, CredentialsSelected, DevProfile, ProfileSteps, QrCodeScanned, Reset},
         user_prompt::CurrentUserPrompt,
-        AppState,
+        AppState, DevMode,
     },
 };
+use log::{debug, info};
 use serde_json::json;
 use std::sync::Arc;
 use uuid::Uuid;
 
-pub(super) async fn create_new_profile(state: AppState) -> Result<AppState, AppError> {
+pub async fn load_dragon_profile(mut state: AppState, dev_profile: DevProfile) -> Result<AppState, AppError> {
+    let steps = dev_profile
+        .execute_step
+        .expect("For dragon profile steps are expected");
+
+    info!("Profile steps executed: {:?}", steps);
+
+    state = reset_settings(state).await?;
+    state = create_new_profile(state).await?;
+
+    if ProfileSteps::AddCredentials <= steps {
+        debug!("Add credentials step executed");
+        state = add_credential(state).await?;
+    }
+
+    if ProfileSteps::AcceptCredentials <= steps {
+        debug!("Accept credentials step executed");
+        state = accept_credential(state).await?;
+    }
+
+    if ProfileSteps::AddConnection <= steps {
+        debug!("Add connection step executed");
+        state = add_connection(state).await?;
+    }
+
+    if ProfileSteps::AcceptConnection <= steps {
+        debug!("Accept connection step executed");
+        state = accept_connection(state).await?;
+    }
+
+    if ProfileSteps::AddPresentation <= steps {
+        debug!("Add presentation step executed");
+        state = add_presentation_request(state).await?;
+    }
+
+    if ProfileSteps::ShareCredentails <= steps {
+        debug!("Share credentials step executed");
+        state = share_credentials(state).await?;
+    }
+
+    if ProfileSteps::AddFutureEngineer <= steps {
+        debug!("Add future engineer step executed");
+        state = add_future_engineer(state).await?;
+    }
+
+    if ProfileSteps::CompleteFlow <= steps {
+        debug!("Accept future engineer step executed");
+        state = accept_future_engineer(state).await?;
+    }
+
+    state.dev_mode = DevMode::OnWithAutologin;
+
+    Ok(state)
+}
+
+
+async fn create_new_profile(state: AppState) -> Result<AppState, AppError> {
     let create_new = CreateNew {
         name: "Shenron".to_string(),
         picture: "&#x1F432".to_string(),
@@ -30,7 +87,7 @@ struct CredentialResponse {
     offer: serde_json::Value,
 }
 
-pub(super) async fn add_credential(state: AppState) -> Result<AppState, AppError> {
+async fn add_credential(state: AppState) -> Result<AppState, AppError> {
     // URL from NGDIL demo
     let url = "https://api.demo.ngdil.com/api/starting-offer";
 
@@ -61,7 +118,7 @@ pub(super) async fn add_credential(state: AppState) -> Result<AppState, AppError
     command::reduce(state, Arc::new(qr_code)).await
 }
 
-pub(super) async fn accept_credential(state: AppState) -> Result<AppState, AppError> {
+async fn accept_credential(state: AppState) -> Result<AppState, AppError> {
     let cr_selected = CredentialOffersSelected {
         offer_indices: vec![0, 1, 2, 3, 4],
     };
@@ -77,7 +134,7 @@ struct ConnectionResponse {
     request_options: serde_json::Value,
 }
 
-pub(super) async fn add_connection(state: AppState) -> Result<AppState, AppError> {
+async fn add_connection(state: AppState) -> Result<AppState, AppError> {
     // URL from NGDIL demo
     let url = "https://api.demo.ngdil.com/siop";
 
@@ -105,7 +162,7 @@ pub(super) async fn add_connection(state: AppState) -> Result<AppState, AppError
     command::reduce(state, Arc::new(qr_code)).await
 }
 
-pub(super) async fn accept_connection(state: AppState) -> Result<AppState, AppError> {
+async fn accept_connection(state: AppState) -> Result<AppState, AppError> {
     command::reduce(state, Arc::new(ConnectionAccepted)).await
 }
 
@@ -117,7 +174,7 @@ struct PresentationResponse {
     request_options: serde_json::Value,
 }
 
-pub(super) async fn add_presentation_request(state: AppState) -> Result<AppState, AppError> {
+async fn add_presentation_request(state: AppState) -> Result<AppState, AppError> {
     // URL from NGDIL demo
     let url = "https://api.demo.ngdil.com/api/oid4vp";
 
@@ -143,7 +200,7 @@ pub(super) async fn add_presentation_request(state: AppState) -> Result<AppState
     command::reduce(state, Arc::new(qr_code)).await
 }
 
-pub(super) async fn share_credentials(state: AppState) -> Result<AppState, AppError> {
+async fn share_credentials(state: AppState) -> Result<AppState, AppError> {
     if let Some(CurrentUserPrompt::ShareCredentials {
         client_name: _,
         logo_uri: _,
@@ -163,7 +220,7 @@ pub(super) async fn share_credentials(state: AppState) -> Result<AppState, AppEr
     }
 }
 
-pub(super) async fn add_future_engineer(state: AppState) -> Result<AppState, AppError> {
+async fn add_future_engineer(state: AppState) -> Result<AppState, AppError> {
     let url = "https://api.demo.ngdil.com/api/credential-offer";
 
     let payload = json!({"credential":"Future Engineer","issuer":"kw1c"});
@@ -185,12 +242,12 @@ pub(super) async fn add_future_engineer(state: AppState) -> Result<AppState, App
     command::reduce(state, Arc::new(qr_code)).await
 }
 
-pub(super) async fn accept_future_engineer(state: AppState) -> Result<AppState, AppError> {
+async fn accept_future_engineer(state: AppState) -> Result<AppState, AppError> {
     let cr_selected = CredentialOffersSelected { offer_indices: vec![0] };
 
     command::reduce(state, Arc::new(cr_selected)).await
 }
 
-pub(super) async fn reset_settings(state: AppState) -> Result<AppState, AppError> {
+async fn reset_settings(state: AppState) -> Result<AppState, AppError> {
     command::reduce(state, Arc::new(Reset)).await
 }
